feat: option to override source of dynamic property label (#19)
* Implement 'labelSource' option for property

Optional string path to the object that should be used as the root context of the property's dynamic label. For example, to evaluate the property label in the parent scope, if you have a property called "Parent", then set `labelSource: 'Parent'`, and when the label is evaluated (ex: for error messages) it will use the value of the 'Parent' property as input to the label format. In other words, "[Name]'s Property" with the labelSource option is equivalent to "[Parent.Name]'s Property" without it.

* Consolidate dynamic label handling in exported functions

* Remove unused import

* Use const